refactor(rest-api): address human review on PR #18537
- UpdatePortalCategoryUseCase: check existence before validating the
  title, so updating a non-existent category returns 404 (not found)
  rather than 400 (validation) when both conditions are true.
- Collapse PortalCategoryDomainService from a core interface + infra
  @Service impl into a single concrete @DomainService class in core,
  per team convention: the interface/infra split is only needed when a
  domain service must reach outside the onion architecture
  (repositories, legacy services), and this one only ever depended on
  the core PortalCategoryCrudService abstraction.
- Moved/renamed its test accordingly to
  core/portal_category/domain_service/PortalCategoryDomainServiceTest,
  mirroring the convention used by other concrete @DomainService
  classes (e.g. DashboardDomainServiceTest).

Left validateTitle() as a blank-check only, no uniqueness check:
duplicate titles are intentionally allowed per the original design -
the story doesn't require category names to be unique.

https://gravitee.atlassian.net/browse/PORTAL-16
