+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/test/java/inmemory/spring/InMemoryConfiguration.java
@@ -32,7 +32,6 @@ import io.gravitee.apim.core.specgen.service_provider.OasProvider;
 import io.gravitee.apim.core.specgen.service_provider.SpecGenNotificationProvider;
 import io.gravitee.apim.core.specgen.service_provider.SpecGenProvider;
 import io.gravitee.apim.core.workflow.query_service.WorkflowQueryService;
-import io.gravitee.apim.infra.domain_service.portal_category.PortalCategoryDomainServiceImpl;
 import io.gravitee.apim.infra.query_service.audit.AuditEventQueryServiceImpl;
 import org.mockito.Mockito;
 import org.springframework.context.annotation.Bean;
@@ -628,6 +627,6 @@ public class InMemoryConfiguration {
 
     @Bean
     public PortalCategoryDomainService portalCategoryDomainService(PortalCategoryCrudServiceInMemory portalCategoryCrudServiceInMemory) {
-        return new PortalCategoryDomainServiceImpl(portalCategoryCrudServiceInMemory);
+        return new PortalCategoryDomainService(portalCategoryCrudServiceInMemory);
     }
 }
